requirements.txt
# FastAPI core
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19

# Database ORM
prisma==0.15.0

# Auth / Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
cryptography==43.0.3

# Stripe
stripe==11.3.0

# HTTP client
httpx==0.28.1

# Config / Env
pydantic[email]>=2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# CORS / Rate limiting
slowapi==0.1.9

# AI
google-genai>=1.17.0

# Retry logic
tenacity==9.0.0

# Timezone
pytz>=2024.1

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
